Replace the "environment" attribute/property with command line "options"
This will help unify the Debian and RHEL providers. The only environment
variable in Debian's defaults file that actually matters is the one for
command line arguments, some of which don't have equivalents that can go
in a config file.

From now on, on all supported platforms, the command line "options" will
be written to `/etc/default/dnsmasq` as the `$DNSMASQ_OPTS` environment
variable and that will be the only thing set in that file.